Moved getTierText
