+++ apps/web/src/utils/pokemon/index.ts
@@ -1,3 +1,2 @@
-export * from './get-tier-name';
 export * from './stats';
 export * from './format-table-data';
